Move Struct to rely less on shared examples
