+++ core/struct/filter_spec.rb
@@ -1,10 +1,8 @@
 require_relative '../../spec_helper'
-require_relative 'shared/select'
-require_relative 'shared/accessor'
-require_relative '../enumerable/shared/enumeratorized'
+require_relative 'fixtures/classes'
 
 describe "Struct#filter" do
-  it_behaves_like :struct_select, :filter
-  it_behaves_like :struct_accessor, :filter
-  it_behaves_like :enumeratorized_with_origin_size, :filter, Struct.new(:foo).new
+  it "is an alias of Struct#select" do
+    StructClasses::Car.instance_method(:filter).should == StructClasses::Car.instance_method(:select)
+  end
 end
