+++ core/struct/inspect_spec.rb
@@ -1,7 +1,8 @@
 require_relative '../../spec_helper'
 require_relative 'fixtures/classes'
-require_relative 'shared/inspect'
 
 describe "Struct#inspect" do
-  it_behaves_like :struct_inspect, :inspect
+  it "is an alias of Struct#to_s" do
+    StructClasses::Car.instance_method(:inspect).should == StructClasses::Car.instance_method(:to_s)
+  end
 end
