+++ core/struct/select_spec.rb
@@ -1,10 +1,31 @@
 require_relative '../../spec_helper'
-require_relative 'shared/select'
+require_relative 'fixtures/classes'
 require_relative 'shared/accessor'
 require_relative '../enumerable/shared/enumeratorized'
 
 describe "Struct#select" do
-  it_behaves_like :struct_select, :select
   it_behaves_like :struct_accessor, :select
   it_behaves_like :enumeratorized_with_origin_size, :select, Struct.new(:foo).new
+
+  it "raises an ArgumentError if given any non-block arguments" do
+    struct = StructClasses::Car.new
+    -> { struct.select(1) { } }.should.raise(ArgumentError)
+  end
+
+  it "returns a new array of elements for which block is true" do
+    struct = StructClasses::Car.new("Toyota", "Tercel", "2000")
+    struct.select { |i| i == "2000" }.should == [ "2000" ]
+  end
+
+  it "returns an instance of Array" do
+    struct = StructClasses::Car.new("Ford", "Escort", "1995")
+    struct.select { true }.should.instance_of?(Array)
+  end
+
+  describe "without block" do
+    it "returns an instance of Enumerator" do
+      struct = Struct.new(:foo).new
+      struct.select.should.instance_of?(Enumerator)
+    end
+  end
 end
