+++ core/struct/values_spec.rb
@@ -2,10 +2,7 @@ require_relative '../../spec_helper'
 require_relative 'fixtures/classes'
 
 describe "Struct#values" do
-  it "is a synonym for to_a" do
-    car = StructClasses::Car.new('Nissan', 'Maxima')
-    car.values.should == car.to_a
-
-    StructClasses::Car.new.values.should == StructClasses::Car.new.to_a
+  it "is an alias of Struct#to_a" do
+    StructClasses::Car.instance_method(:values).should == StructClasses::Car.instance_method(:to_a)
   end
 end
